requirements.txt
google-genai
duckduckgo-search
requests
python-dotenv
